Compile fix

+++ src/ui/dialogs/lua_editor_dlg.cpp
@@ -116,7 +116,8 @@ LuaEditorDlg::LuaEditorDlg(
                     // the same indent as the line it's attached to.
                     // Also, add 3; this is just a hack to account for the width of the box, there
                     // doesn't seem to be any way to get it directly from Scintilla.
-                    const int indent = editor->GetLineIndentation(lineNumber) + FromDIP(3);
+                    const unsigned long indent =
+                        editor->GetLineIndentation(lineNumber) + FromDIP(3);
 
                     const int widthAnn =
                         editor->TextWidth(Wisteria::UI::CodeEditor::ERROR_ANNOTATION_STYLE,
